+++ dfm_tools/observations.py
@@ -20,9 +20,11 @@ import glob
 import matplotlib.pyplot as plt
 import matplotlib.dates as md
 import shutil
+import fiona
 
 __all__ = ["ssh_catalog_subset",
            "ssh_catalog_toxynfile",
+           "ssh_catalog_tokmlfile",
            "ssh_retrieve_data",
            "ssh_netcdf_overview",
            ]
@@ -943,6 +945,9 @@ def ssh_retrieve_data(ssh_catalog_gpd, dir_output, time_min=None, time_max=None,
 
 
 def ssh_catalog_toxynfile(ssc_catalog_gpd, file_xyn):
+    """
+    converts a ssh_catalog to a _obs.xyn file for easy visualisation in FM GUI and interacter
+    """
     lon = ssc_catalog_gpd.geometry.x
     lat = ssc_catalog_gpd.geometry.y
     name = ssc_catalog_gpd['station_name_unique']
@@ -950,8 +955,31 @@ def ssh_catalog_toxynfile(ssc_catalog_gpd, file_xyn):
     np.savetxt(file_xyn, data, fmt='%13.8f %13.8f %-s')
 
 
-def ssh_netcdf_overview(dir_netcdf, perplot=30, time_min=None, time_max=None, yearstep=None):
+def ssh_catalog_tokmlfile(ssc_catalog_gpd, file_kml):
+    """
+    converts a ssh_catalog to a kml file for easy visualisation in google earth
+    """
+    # Enable fiona driver
+    # TODO: gpd sometimes fails with "AttributeError: 'NoneType' object has no attribute 'drvsupport'" 
+    # gpd.io.file.fiona.drvsupport.supported_drivers['KML'] = 'rw' # 
+    # gpd.io.file.fiona.drvsupport.supported_drivers['LIBKML'] = 'rw'
+    fiona.supported_drivers['KML'] = 'rw'
+    fiona.supported_drivers['LIBKML'] = 'rw'
     
+    #select only names and geometry column to reduce file size
+    ssc_catalog_gpd_minimal = ssc_catalog_gpd[["station_name_unique","geometry"]]
+    # rename to "name" results in a label in Google Earth
+    ssc_catalog_gpd_minimal = ssc_catalog_gpd_minimal.rename({"station_name_unique":"name"},axis=1)
+    
+    # Write file
+    ssc_catalog_gpd_minimal.to_file(file_kml, driver='KML')
+
+
+def ssh_netcdf_overview(dir_netcdf, perplot=30, time_min=None, time_max=None, yearstep=None):
+    """
+    reads all netcdf files in a directory and makes figures of the non-nan waterlevel time availability
+    it also writes a csv file with statistics
+    """
     dir_output = os.path.join(dir_netcdf, "overview")
     if os.path.isdir(dir_output):
         shutil.rmtree(dir_output)
@@ -1046,9 +1074,3 @@ def ssh_netcdf_overview(dir_netcdf, perplot=30, time_min=None, time_max=None, ye
     file_csv = os.path.join(dir_output, "waterlevel_data_netcdf_overview.csv")
     stats.to_csv(file_csv, float_format="%.2f")
     
-    # write xynfile
-    file_xyn = os.path.join(dir_output, "stations_obs.xyn")
-    lon, lat, name = stats["longitude"], stats["latitude"], stats["fname_clean"]
-    data = np.c_[lon, lat, name]
-    np.savetxt(file_xyn, data, fmt='%13.8f %13.8f %-s')
-